refactor(checkbox): adjusting checkbox component after review
