+++ packages/yoga/src/Checkbox/CheckIcon.js
